+++ lib/src/api/channels/partial_text_channel.dart
@@ -10,13 +10,13 @@ import 'package:mineral/src/internal/mixins/mineral_client.dart';
 import 'package:mineral_ioc/ioc.dart';
 
 class PartialTextChannel extends GuildChannel{
-  final MessageManager _messages;
+  final MessageManager<Message> _messages;
   final Snowflake? _lastMessageId;
 
   PartialTextChannel(this._messages, this._lastMessageId, super.guildId, super.parentId, super.label, super.type, super.position, super.flags, super.permissions, super.id);
 
   /// Get [MessageManager]
-  MessageManager get messages => _messages;
+  MessageManager<Message> get messages => _messages;
 
   /// Get last [Message] of this
   Message? get lastMessage => _messages.cache.get(_lastMessageId);