+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.6.0
 - Remove mixins to public access
 - Add `createdAt` and `updatedAt` to `Message`
+- Add correct message type from `fetch()`
 
 ## 2.5.0
 - Add `make:service`